pyproject.toml
[tool.poetry]
name = "attention"
version = "0.1.0"
description = ""
authors = ["Thomas Sajot"]

packages = [
    { include = "attention" , from ="src"},
]

[tool.poetry.dependencies]
python = "3.10.*"
jax = "0.3.*"
jaxlib = "*"
numpy = "*"
dm-haiku = "*"
torch = "*"
torchvision = "*"
jupyterlab = "*"
mediapy = "*"
tqdm = "*"
ipywidgets = "*"


[tool.poetry.dev-dependencies]
pytest = "7.*"
flake8 = "4.0.*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
